requirements.txt
typing>=3.7.4.3
httpx==0.24.1
python-dateutil>=2.8.2
pydantic>=2.0.2